+++ src/Modbus/ModbusSlave.h
@@ -191,11 +191,11 @@ public:
 //-------------------------------------------------------------------------------
     std::string m_sModbusSlaveLinkLayerName;
     CModbusSlaveLinkLayerInterface* m_pxModbusSlaveLinkLayer;
-    uint8_t m_uiModbusSlaveLinkLayerId;
+    uint16_t m_uiModbusSlaveLinkLayerId;
 
     std::string m_sDeviceControlName;
     CDeviceControl* m_pxDeviceControl;
-    uint8_t m_uiDeviceControlId;
+    uint16_t m_uiDeviceControlId;
 
     uint8_t m_uiOwnAddress;
     uint8_t m_uiSlaveAddress;
