+++ src/SettingsLoad.h
@@ -64,13 +64,13 @@ private:
 
 private:
     std::string m_sInternalModuleName;
-    uint8_t m_uiInternalModuleId;
+    uint16_t m_uiInternalModuleId;
 
     std::string m_sDeviceControlName;
-    uint8_t m_uiDeviceControlId;
+    uint16_t m_uiDeviceControlId;
 
     std::string m_sDataStoreName;
-    uint8_t m_uiDataStoreId;
+    uint16_t m_uiDataStoreId;
 
     uint8_t* m_puiIntermediateBuff;
